+++ bioframe/__init__.py
@@ -1,9 +1,8 @@
 #-*- coding: utf-8 -*-
 from __future__ import division, print_function
 from ._version import __version__
-from . import io
-from . import tools
-from .io.formats import *
-from .io.resources import *
-from .core import *
+from .region import *
+from .frameops import *
+from .formats import *
+from .resources import *
 from .schemas import SCHEMAS
