+++ bioframe/_version.py
@@ -1,1 +1,1 @@
-__version__ = '0.0.9'
+__version__ = '0.1.0-dev'
