--- bioframe/io/__init__.py
@@ -1,5 +1,0 @@
-#-*- coding: utf-8 -*-
-from __future__ import division, print_function
-from . import formats
-from . import resources
-from . import process
